# frontend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OPNGMS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/tenants/{tenant_id}/reports": {
      "post": {
        "tags": [
          "reports"
        ],
        "summary": "Generate Report",
        "operationId": "generate_report_api_tenants__tenant_id__reports_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "List Generated Reports",
        "operationId": "list_generated_reports_api_tenants__tenant_id__reports_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GeneratedReportOut"
                  },
                  "title": "Response List Generated Reports Api Tenants  Tenant Id  Reports Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenant_id}/reports/{report_id}/download": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Download Generated Report",
        "operationId": "download_generated_report_api_tenants__tenant_id__reports__report_id__download_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "report_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Report Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenant_id}/reports/languages": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report Languages",
        "operationId": "get_report_languages_api_tenants__tenant_id__reports_languages_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ReportLanguageOut"
                  },
                  "title": "Response Get Report Languages Api Tenants  Tenant Id  Reports Languages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenant_id}/reports/settings": {
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report Settings",
        "operationId": "get_report_settings_api_tenants__tenant_id__reports_settings_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "reports"
        ],
        "summary": "Update Report Settings",
        "operationId": "update_report_settings_api_tenants__tenant_id__reports_settings_put",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReportSettingsIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/tenants/{tenant_id}/reports/settings/logo": {
      "put": {
        "tags": [
          "reports"
        ],
        "summary": "Upload Report Logo",
        "operationId": "upload_report_logo_api_tenants__tenant_id__reports_settings_logo_put",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_report_logo_api_tenants__tenant_id__reports_settings_logo_put"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "reports"
        ],
        "summary": "Delete Report Logo",
        "operationId": "delete_report_logo_api_tenants__tenant_id__reports_settings_logo_delete",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportSettingsOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "reports"
        ],
        "summary": "Get Report Logo",
        "operationId": "get_report_logo_api_tenants__tenant_id__reports_settings_logo_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_report_logo_api_tenants__tenant_id__reports_settings_logo_put": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_report_logo_api_tenants__tenant_id__reports_settings_logo_put"
      },
...
      "GeneratedReportOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "period_from": {
            "type": "string",
            "format": "date-time",
            "title": "Period From"
          },
          "period_to": {
            "type": "string",
            "format": "date-time",
            "title": "Period To"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "size": {
            "type": "integer",
            "title": "Size"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "kind",
          "period_from",
          "period_to",
          "created_by",
          "size",
          "created_at"
        ],
        "title": "GeneratedReportOut"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReportLanguageOut": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "ReportLanguageOut"
      },
      "ReportRequest": {
        "properties": {
          "from": {
            "type": "string",
            "format": "date-time",
            "title": "From"
          },
          "to": {
            "type": "string",
            "format": "date-time",
            "title": "To"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          }
        },
        "type": "object",
        "required": [
          "from",
          "to"
        ],
        "title": "ReportRequest"
      },
...
      "ReportSettingsIn": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "owner": {
            "type": "string",
            "title": "Owner",
            "default": ""
          },
          "timezone": {
            "type": "string",
            "title": "Timezone",
            "default": "UTC"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "from_email": {
            "type": "string",
            "title": "From Email",
            "default": ""
          },
          "sections": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Sections",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "title"
        ],
        "title": "ReportSettingsIn"
      },
      "ReportSettingsOut": {
        "properties": {
          "title": {
            "type": "string",
            "title": "Title"
          },
          "owner": {
            "type": "string",
            "title": "Owner"
          },
          "timezone": {
            "type": "string",
            "title": "Timezone"
          },
          "has_logo": {
            "type": "boolean",
            "title": "Has Logo"
          },
          "logo_mime": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Logo Mime"
          },
          "language": {
            "type": "string",
            "title": "Language"
          },
          "from_email": {
            "type": "string",
            "title": "From Email"
          },
          "sections": {
            "additionalProperties": {
              "type": "boolean"
            },
            "type": "object",
            "title": "Sections",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "title",
          "owner",
          "timezone",
          "has_logo",
          "logo_mime",
          "language",
          "from_email"
        ],
        "title": "ReportSettingsOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}